:green_heart: :extract: Extract test files from parser_test.go
