+++ parser/parser_test.go
@@ -1,11 +1,9 @@
 package parser
 
 import (
-	"strings"
 	"testing"
 
 	"github.com/akm/pas/ast"
-	"github.com/stretchr/testify/assert"
 )
 
 func TestParseEmptyUnit(t *testing.T) {
@@ -461,52 +459,4 @@ func TestComments(t *testing.T) {
 				{Kind: ast.ImplementationSection},
 			},
 		})
-}
-
-func Test2RecordsInATypeBlock(t *testing.T) {
-	parseFile(t, `
-  unit U;
-  interface
-  type
-    T1 = record
-	  A: Integer;
-    end;
-    T2 = record
-    end;
-  implementation
-  {$R *.dfm}
-  end.`,
-		&ast.File{
-			Kind: ast.Unit,
-			Name: "U",
-			Sections: []*ast.FileSection{
-				{
-					Kind: ast.InterfaceSection,
-					Blocks: []ast.FileSectionBlock{
-						ast.TypeBlock{
-							&ast.Record{
-								Name: "T1",
-								Members: []ast.ClassMember{
-									&ast.Variable{Name: "A", Type: "Integer"},
-								},
-							},
-							&ast.Record{
-								Name: "T2",
-							},
-						},
-					},
-				},
-				{Kind: ast.ImplementationSection},
-			},
-		})
-}
-
-func parseFile(t *testing.T, code string, want *ast.File) {
-	t.Helper()
-	code = strings.Replace(code, "\n", "\r\n", -1)
-	f, err := ParseString(code)
-	if err != nil {
-		t.Fatalf("%+v\n", err)
-	}
-	assert.Equal(t, want, f)
 }
